+++ persistence/src/test/java/dao/TireDaoImplTest.java
@@ -48,8 +48,8 @@ public class TireDaoImplTest extends AbstractTestNGSpringContextTests {
         tire1.setPrice(new BigDecimal(45000));
         tire2.setPrice(new BigDecimal(4567));
 
-        tire1.setSize(340);
-        tire2.setSize(340);
+        tire1.setSize(BigDecimal.valueOf(340));
+        tire2.setSize(BigDecimal.valueOf(340));
 
         tireDao.create(tire1);
         tireDao.create(tire2);
